requirements.txt
biopython==1.81
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0
